requirements.txt
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
expects==0.9.0
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
Markdown==3.3.6
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytest-responses==0.5.0
pytest-sugar==0.9.4
pytz==2021.3
requests==2.27.1
responses==0.16.0
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
toml==0.10.2
urllib3==1.26.8
zipp==3.7.0
